Add files via upload
The airport charges $4.00 per hour (first hour) and additional $3 per hour for each hour or part thereof. The maximum charge is $40.00 per day. Assume maximum number of parking hours is 24.

A program that calculates and displays the parking charges for a customer is required. 

here is  a C# program that prompts the user to choose either Long-Term Parking or Short-Term Parking.

-	If the user chooses a Long-Term Parking, the program should prompt the user to enter the number of days, then calculates and displays the parking charge

-	If the user chooses a Short-Term Parking, the program should prompt the user to enter the hours parked for the customer, then calculates and display the parking charge. Consider the requirements listed before.