Add files via upload
A dental Equipment Company pays its sales employees on a commission basis.
The monthly salary of the sales employee is $1000 plus 10% of their gross sales for that month. For example, a salesperson who sells $10000 worth of Diamond instruments for dentists in a month receives 1000 plus 10% of $10000, or a total of $2000. 

here is  a program that uses a while statement to input each salesperson’s gross sales for the past 12 month and calculates and displays that salesperson’s earnings per month and show the total at the end as well. To stop the program the user type -1 and display the message "Goodbye!".